Remove utils and modify the other modules to work with after this change
