+++ simulator/modules/clients.py
@@ -1,4 +1,4 @@
-from utils import get_logger
+from log import get_logger
 import simpy
 
 class Request(object):
