+++ simulator/modules/garbage.py
@@ -1,4 +1,4 @@
-from utils import get_logger
+from log import get_logger
 
 class GC(object):
