+++ tests/context.py
@@ -3,4 +3,4 @@ import sys
 sys.path.insert(0, os.path.abspath('..'))
 
 from simulator.modules import Clients, GC, GCI, Request, LoadBalancer, Server, ServerWithGCI
-from utils import get_config
+from config import get_config
